third_party/WebKit/LayoutTests/external/wpt/fetch/api/request/request-init-003.sub-expected.txt

This is a testharness.js-based test.
FAIL Check request values when initialized from Request assert_equals: Check cache attribute expected (string) "no-cache" but got (undefined) undefined
FAIL Check request values when initialized from Request and init values assert_equals: Check cache attribute expected (string) "no-cache" but got (undefined) undefined
FAIL Check request values when initialized from url string assert_equals: Check url attribute expected "http://url.test:1234/path/subpath?query=true#fragment" but got "http://url.test:1234/path/subpath?query=true"
FAIL Check request values when initialized from url and init values assert_equals: Check cache attribute expected (string) "no-cache" but got (undefined) undefined
Harness: the test ran to completion.
